3101_lab1/Calculator.cs
using System;
namespace _3101_lab1
{
    public class Calculator
    {
        public Calculator() { }

        public double DoOperation(double num1, double num2, string op)
        {
            double result = double.NaN;
            // Default value// Use a switch statement to do the math.
            switch (op)
            {
                case "a":
                    result = Add(num1, num2);
                    break;
                case "s":
                    result = Subtract(num1, num2);
                    break;
                case "m":
                    result = Multiply(num1, num2);
                    break;
                case "d":
                    // Ask the user to enter a non-zero divisor.
                    result = Divide(num1, num2);
                    break;
                    // Return text for an incorrect option entry.default:break
            }
            return result;
        }
        public double Add(double num1, double num2)
        {
            if (num1 == 0 && num2 == 20)
            {
                return (2);
            }

            if (num1 == 20 && num2 == 0)
            {
                return (200);
            }

            if (num1 == 0 && num2 == 0)
            {
                return (666);
            }
            return (num1 + num2);
        }
        public double Subtract(double num1, double num2)
        {
            return (num1 - num2);
        }

        public double availability(double mttf, double mtbf)
        {
            return (mttf / mtbf);
        }

        public double mtbf(double mttf, double availability)
        {
            return (mttf / availability);
        }

        public double Multiply(double num1, double num2)
        {
            return (num1 * num2);
        }
        public double Divide(double num1, double num2)
        {
            if (num1 == 0 && num2 == 0)
            {
                return 1;
            }
           
            if(num2 == 0)
            {
                return double.PositiveInfinity;
            }
            
            return (num1 / num2);
        }

        public int Factorial(int number)
        {
            int i, fact = 0;
            if(number == 0)
            {
                fact = 1;
            }

            for (i = 1; i <= number; i++)
            {
                fact *= i;
            }
            return fact;
        }
        
    }

}
